+++ app/src/main/java/com/emotionstorage/emotionstorage/ui/AppNavHost.kt
@@ -20,13 +20,13 @@ import com.emotionstorage.auth.ui.SignupCompleteScreen
 import com.emotionstorage.daily_report.ui.DailyReportDetailScreen
 import com.emotionstorage.domain.model.User.AuthProvider
 import com.emotionstorage.home.ui.HomeScreen
-import com.emotionstorage.my.ui.AccountInfoScreen
-import com.emotionstorage.my.ui.KeyDescriptionScreen
-import com.emotionstorage.my.ui.MyPageScreen
+import com.emotionstorage.my.ui.accountInfo.AccountInfoScreen
+import com.emotionstorage.my.ui.keyDescription.KeyDescriptionScreen
+import com.emotionstorage.my.ui.myPage.MyPageScreen
 import com.emotionstorage.my.ui.NicknameChangeScreen
-import com.emotionstorage.my.ui.NotificationSettingScreen
+import com.emotionstorage.my.ui.notificationSettings.NotificationSettingScreen
 import com.emotionstorage.my.ui.TermsAndPrivacyScreen
-import com.emotionstorage.my.ui.WithDrawNoticeScreen
+import com.emotionstorage.my.ui.withdraw.WithDrawNoticeScreen
 import com.emotionstorage.time_capsule.ui.ArrivedTimeCapsulesScreen
 import com.emotionstorage.time_capsule.ui.CalendarScreen
 import com.emotionstorage.time_capsule.ui.FavoriteTimeCapsulesScreen
